requirements.txt
discord.py==0.14.3
asyncio
fbchat
PyNaCl
